Init Project con vite, tailwind css y nueva arquitectura Co-authored-by: MarkoEv markodevp@gmail.com
