--- .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,40 +1,0 @@
----
-
-name: "🐛 Bug Report"
-about: "Reportar un error para ayudarnos a mejorar el proyecto."
-title: "[BUG] - Descripción corta del error"
-labels: bug
-assignees: ""
-
----
-
-## 🐛 Bug Report
-
-### Descripción
-
-Describe claramente el error que has encontrado. Incluye detalles sobre cómo se presenta el problema.
-
-### Pasos para reproducir
-
-1. Ir a '...'
-2. Hacer clic en '...'
-3. Describir cualquier otra acción hasta que ocurra el problema.
-4. Error que aparece: '...'
-
-### Comportamiento esperado
-
-Describe lo que esperabas que ocurriera al realizar los pasos anteriores.
-
-### Capturas de pantalla o registros
-
-Si es posible, añade capturas de pantalla o registros de errores.
-
-### Entorno
-
-- Sistema operativo: [p. ej., Windows 10, macOS Catalina]
-- Navegador y versión: [p. ej., Chrome 87, Safari 14]
-- Versión del proyecto: [p. ej., 1.0.0]
-
-### Información adicional
-
-Cualquier otra información relevante.
