+++ package.json
@@ -1,0 +1,19 @@
+{
+  "name": "generador-de-curp",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
+  },
+  "devDependencies": {
+    "vite": "^8.0.0"
+  },
+  "dependencies": {
+    "@tailwindcss/vite": "^4.2.1",
+    "tailwind": "^4.0.0",
+    "tailwindcss": "^4.2.1"
+  }
+}
