+++ src/main.js
@@ -1,0 +1,15 @@
+import "./style.css";
+
+import {
+  generarEstados,
+  generarDays,
+  generarMonths,
+  generarYears,
+} from "./generadores.js";
+
+window.onload = () => {
+  generarEstados();
+  generarDays();
+  generarMonths();
+  generarYears();
+};
